+++ assets/javascripts/kitten/components/comments/comment/stories.js
@@ -1,9 +1,11 @@
 import React from 'react'
+import styled from 'styled-components'
 import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
 import { withKnobs, text, object, boolean } from '@storybook/addon-knobs'
 import { Comment } from './index'
 import { Grid, GridCol } from '../../../components/grid/grid'
+import { CheckedCircleIcon } from '../../../components/icons/checked-circle-icon'
 
 const info = {
   text: `
@@ -47,6 +49,11 @@ const info = {
     ~~~js
     <Comment text="…" />
     ~~~
+
+    ### With avatarBadge
+    ~~~js
+    <Comment avatarBadge="…" />
+    ~~~
   `,
   header: false,
   propTables: false,
@@ -70,6 +77,19 @@ storiesOf('Comments/Comment', module)
               'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris',
             )}
             ownerName={text('Owner name', 'Lorem ipsum')}
+            avatarBadge={
+              boolean('Avatar badge', false) && (
+                <span aria-label="Owner" role="aside">
+                  <CheckedCircleIcon
+                    width="25"
+                    height="25"
+                    circleColor="#19b4fa"
+                    checkedColor="#fff"
+                    aria-hidden="true"
+                  />
+                </span>
+              )
+            }
             bottomNotes={
               boolean('With bottom notes?', false) && (
                 <div>