+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Update `PasswordInput` with styled-components.
+- Feature: Add `title` prop on `CheckedCircleIcon`.
+- Feature: Add `avatarBadge` prop on `CommentAvatar`, `Comment` and `CommentForm`.
 
 ## [2.15.0] - 2019-07-16